Issue #2336: Extract assertSuccessfulRequestReturnsResponseBody test function.
